Add link in login error message
