+++ plugins/org.jboss.tools.openshift.common.ui/src/org/jboss/tools/openshift/internal/common/ui/explorer/OpenShiftExplorerView.java
@@ -52,7 +52,7 @@ import org.jboss.tools.openshift.internal.common.ui.connection.ConnectionWizard;
  * @author Xavier Coulon
  * @author Andre Dietisheim
  */
-public class OpenShiftExplorerView extends CommonNavigator implements IConnectionsRegistryListener {
+public class OpenShiftExplorerView extends BaseOpenShiftExplorerView implements IConnectionsRegistryListener {
 
 	private Control connectionsPane;
 	private Control explanationsPane;
@@ -214,68 +214,6 @@ public class OpenShiftExplorerView extends CommonNavigator implements IConnectio
 					// }
 				}
 			};
-		}
-	}
-
-	@Override
-	public void initListeners(TreeViewer viewer) {
-		super.initListeners(viewer);
-		Tree tree = viewer.getTree();
-		if (tree != null && !tree.isDisposed()) {
-			new LinkMouseListener(tree);
-		}
-	}
-
-	static class LinkMouseListener extends MouseAdapter implements MouseMoveListener {
-		Tree tree;
-
-		LinkMouseListener(Tree tree) {
-			this.tree = tree;
-			tree.addMouseListener(this);
-			tree.addMouseMoveListener(this);
-		}
-
-		boolean isLink = false;
-
-		@Override
-		public void mouseMove(MouseEvent e) {
-			if (tree.isDisposed()) {
-				return;
-			}
-			ILink link = getLink(e);
-			if (isLink != (link != null)) {
-				isLink = (link != null);
-				Cursor cursor = isLink ? Display.getDefault().getSystemCursor(SWT.CURSOR_HAND) : null;
-				tree.setCursor(cursor);
-			}
-		}
-
-		@Override
-		public void mouseUp(MouseEvent e) {
-			final ILink link = getLink(e);
-			if (link != null) {
-				Display.getDefault().asyncExec(new Runnable() {
-					@Override
-					public void run() {
-						if (tree.isDisposed()) {
-							return;
-						}
-						tree.setCursor(null);
-						isLink = false;
-						link.execute();
-					}
-				});
-			}
-		}
-
-		ILink getLink(MouseEvent e) {
-			if (e.getSource() instanceof Tree) {
-				Tree tree = (Tree) e.getSource();
-				TreeItem t = tree.getItem(new Point(e.x, e.y));
-				Object o = t == null ? null : t.getData();
-				return o instanceof ILink ? (ILink) o : null;
-			}
-			return null;
 		}
 	}
