+++ plugins/org.jboss.tools.openshift.ui/src/org/jboss/tools/openshift/internal/ui/applicationexplorer/OpenShiftApplicationExplorerLabelProvider.java
@@ -11,7 +11,10 @@
 package org.jboss.tools.openshift.internal.ui.applicationexplorer;
 
 import org.eclipse.jface.viewers.StyledString;
+import org.eclipse.jface.viewers.StyledString.Styler;
+import org.eclipse.swt.SWT;
 import org.eclipse.swt.graphics.Image;
+import org.eclipse.swt.graphics.TextStyle;
 import org.jboss.tools.openshift.core.odo.KubernetesLabels;
 import org.jboss.tools.openshift.internal.common.ui.OpenShiftCommonImages;
 import org.jboss.tools.openshift.internal.common.ui.explorer.BaseExplorerLabelProvider;
@@ -19,6 +22,7 @@ import org.jboss.tools.openshift.internal.ui.OpenShiftImages;
 import org.jboss.tools.openshift.internal.ui.models.applicationexplorer.ApplicationElement;
 import org.jboss.tools.openshift.internal.ui.models.applicationexplorer.ApplicationExplorerUIModel;
 import org.jboss.tools.openshift.internal.ui.models.applicationexplorer.ComponentElement;
+import org.jboss.tools.openshift.internal.ui.models.applicationexplorer.MessageElement;
 import org.jboss.tools.openshift.internal.ui.models.applicationexplorer.ProjectElement;
 import org.jboss.tools.openshift.internal.ui.models.applicationexplorer.ServiceElement;
 import org.jboss.tools.openshift.internal.ui.models.applicationexplorer.StorageElement;
@@ -54,8 +58,23 @@ public class OpenShiftApplicationExplorerLabelProvider extends BaseExplorerLabel
 			return style(((URLElement)element).getWrapped().getName(),String.join(" ", ((URLElement)element).getWrapped().getPort(), ((URLElement)element).getWrapped().getState().toString()), limit);
 		} else if (element instanceof ServiceElement) {
 			return style(KubernetesLabels.getComponentName(((ServiceElement)element).getWrapped()), "", limit);
+		} else if (element instanceof MessageElement) {
+			return getStyledText((MessageElement) element);
 		}
 		return super.getStyledText(element, limit);
+	}
+	
+	private StyledString getStyledText(MessageElement messageElement) {
+		StyledString styledString = new StyledString();
+		styledString.append(messageElement.getWrapped(), new Styler() {
+			
+			@Override
+			public void applyStyles(TextStyle textStyle) {
+				textStyle.underline = true;
+				textStyle.underlineStyle = SWT.UNDERLINE_LINK;
+			}
+		});
+		return styledString;
 	}
 
 	@Override
